backend/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
langchain==0.3.7
langchain-community==0.3.7
chromadb==0.5.20
sentence-transformers==3.2.1
pyyaml==6.0.2
httpx==0.27.2
pydantic==2.9.2
google-generativeai==0.8.3
